+add HealingEffect, BloodEffect, Bug Fix
+ Fixed some bugs with player attack

+++ Assets/[GAME]/Scripts/Interfaces/IHealable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IHealable
 {
-    void Heal(int healAmount);
+    void Heal(float healAmount);
 }
